+++ Consultoresvs3/Models/Proyecto.cs
@@ -14,14 +14,14 @@ namespace Consultoresvs3.Models
         public string Nombre { get; set; }
 
         [Display(Name = "Honorario del Proyecto")]
-        [DisplayFormat(DataFormatString = "{0:C2}", ApplyFormatInEditMode = false)]
+        [DisplayFormat(DataFormatString = "{0:C2}",ApplyFormatInEditMode = false)]
         public decimal Precio { get; set; }
 
         [Display(Name = "Tiempo Estimado (Horas)")]
-        public decimal TiempoEstipulado { get; set; }
+        public int TiempoEstipulado { get; set; }
         //Horas totales trabajadas
         [Display(Name = "Horas Trabajadas")]
-        public decimal HorasTrabajdas { get; set; }
+        public int HorasTrabajdas { get; set; }
         //Empresa
         [Display(Name = "Empresa")]
         public int IdEmpresa { get; set; }
